evez-ecosystem/evezart-repos/evez-atlas/requirements.txt
# Quantum Threat Detection System Dependencies
# Core dependencies for the causal chain API and tooling.

fastapi>=0.110
httpx>=0.26
pydantic>=2.6
python-dotenv>=1.0
slowapi>=0.1.9
uvicorn>=0.27

# Testing utilities
pytest>=8.0
pytest-asyncio>=0.23

# Semantic Possibility Space dependencies
sentence-transformers==3.0.1
torch==2.1.2
numpy==1.26.4

# Optional: For production use with real quantum hardware:
# qiskit>=1.0
# qiskit-machine-learning
# qiskit-ibm-runtime

# Optional: For data processing:
# pandas
# scikit-learn

# Optional: For visualization:
# matplotlib

# Bionetwork optimizer dependencies
torch_geometric>=2.5
networkx>=3.2
